+++ src/deluge/model/global_effectable/global_effectable_for_clip.cpp
@@ -178,7 +178,7 @@ doNormal:
 			                             &delayWorkingState, analogDelaySaturationAmount);
 			processStutter(globalEffectableBuffer, numSamples, paramManagerForClip);
 
-			int32_t postReverbSendVolumeIncrement = (int32_t)(postReverbVolume - postReverbVolumeLastTime) / numSamples;
+			int32_t postReverbSendVolumeIncrement = (int32_t)((float)(postReverbVolume - postReverbVolumeLastTime) / (float)numSamples);
 
 			processReverbSendAndVolume(globalEffectableBuffer, numSamples, reverbBuffer, volumePostFX,
 			                           postReverbVolumeLastTime, reverbSendAmount, pan, true,
